requirements.txt
# Freedom Wallet Bot - Python Dependencies

# Core bot framework
python-telegram-bot==20.7
telegram==0.0.1

# AI/NLP
openai==1.10.0
anthropic==0.18.0  # Claude API (alternative)

# Vector database (knowledge base)
chromadb==0.4.22
sentence-transformers==2.3.1

# Image processing (for personalized celebrations)
Pillow==10.2.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL
alembic==1.13.1  # Database migrations

# Google Sheets integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==5.12.3
oauth2client==4.1.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
aiohttp==3.9.1

# Logging & monitoring
loguru==0.7.2
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Production server
gunicorn==21.2.0
uvicorn==0.27.0

# Caching
redis==5.0.1
